requirements.txt
Twisted==15.0.0
numpy==1.9.2
opentsdb-pandas==0.0.1
pandas==0.14.1
python-dateutil==2.4.1
pytz==2015.2
requests==2.6.0
six==1.9.0
ujson==1.33
wsgiref==0.1.2
zope.interface==4.1.2
